order_service/requirements.txt
Django==4.2
djangorestframework==3.15
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3
requests==2.31.0
python-dotenv==1.0.1
